♻️ move pool reserves derivation to services/pools
