+++ web/src/hooks/usePoolReserves.ts
@@ -1,8 +1,8 @@
 import { useMemo } from "react";
 import { useObservable } from "react-rx";
 import { map } from "rxjs";
+import { getAssetHubPoolReserves$ } from "../services/pools/reserves";
 import type { Pool } from "../services/pools/types";
-import { getAssetHubPoolReserves$ } from "../state/pools";
 
 type UsePoolReservesProps = {
 	pool: Pool | null | undefined;
